Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -19,7 +19,7 @@ int main(){
         printf("%d",result);
         break;
         case'/':
-        result=a/b;
+        result=float(a/b);
         printf("%d",result);
         break;
         default:
